# 04-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinanceIntel Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/imports/upload": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Upload Csv",
        "description": "Step 1: Upload CSV, create ImportBatch, insert ImportRows.",
        "operationId": "upload_csv_api_v1_imports_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_csv_api_v1_imports_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/imports/{batch_id}/commit": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Commit Import Batch",
        "description": "Step 2: Commit batch using column_mapping.\nCalls the commit_batch service.",
        "operationId": "commit_import_batch_api_v1_imports__batch_id__commit_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/{batch_id}": {
      "delete": {
        "tags": [
          "imports"
        ],
        "summary": "Delete Import Batch",
        "operationId": "delete_import_batch_api_v1_imports__batch_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Import Batch Api V1 Imports  Batch Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaign-runs/": {
      "get": {
        "tags": [
          "campaign-runs"
        ],
        "summary": "List Campaign Runs",
        "operationId": "list_campaign_runs_api_v1_campaign_runs__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampaignRunOut"
                  },
                  "type": "array",
                  "title": "Response List Campaign Runs Api V1 Campaign Runs  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "campaign-runs"
        ],
        "summary": "Create Campaign Run",
        "operationId": "create_campaign_run_api_v1_campaign_runs__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/campaign-runs/{run_id}": {
      "get": {
        "tags": [
          "campaign-runs"
        ],
        "summary": "Get Campaign Run",
        "operationId": "get_campaign_run_api_v1_campaign_runs__run_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "campaign-runs"
        ],
        "summary": "Update Campaign Run",
        "operationId": "update_campaign_run_api_v1_campaign_runs__run_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignRunUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaign-run-stats/upsert": {
      "post": {
        "tags": [
          "campaign-run-stats"
        ],
        "summary": "Upsert Campaign Run Stat",
        "operationId": "upsert_campaign_run_stat_api_v1_campaign_run_stats_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignRunStatCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunStatOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/campaign-run-stats/": {
      "get": {
        "tags": [
          "campaign-run-stats"
        ],
        "summary": "List Campaign Run Stats",
        "operationId": "list_campaign_run_stats_api_v1_campaign_run_stats__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignRunStatOut"
                  },
                  "title": "Response List Campaign Run Stats Api V1 Campaign Run Stats  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_csv_api_v1_imports_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_csv_api_v1_imports_upload_post"
      },
      "CampaignRunCreate": {
        "properties": {
          "campaign_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "ad_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Account Id"
          },
          "buyer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Buyer Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "buyer_id",
          "started_at"
        ],
        "title": "CampaignRunCreate"
      },
      "CampaignRunOut": {
        "properties": {
          "campaign_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "ad_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Account Id"
          },
          "buyer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Buyer Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          }
        },
        "type": "object",
        "required": [
          "buyer_id",
          "started_at",
          "id",
          "company_id"
        ],
        "title": "CampaignRunOut"
      },
      "CampaignRunStatCreate": {
        "properties": {
          "campaign_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Run Id"
          },
          "stat_date": {
            "type": "string",
            "format": "date",
            "title": "Stat Date"
          },
          "spend": {
            "type": "string",
            "format": "decimal",
            "title": "Spend",
            "default": "0.00"
          },
          "revenue": {
            "type": "string",
            "format": "decimal",
            "title": "Revenue",
            "default": "0.00"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "fx_rate_to_base": {
            "type": "string",
            "format": "decimal",
            "title": "Fx Rate To Base"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          }
        },
        "type": "object",
        "required": [
          "campaign_run_id",
          "stat_date",
          "currency",
          "fx_rate_to_base",
          "source"
        ],
        "title": "CampaignRunStatCreate"
      },
      "CampaignRunStatOut": {
        "properties": {
          "campaign_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Run Id"
          },
          "stat_date": {
            "type": "string",
            "format": "date",
            "title": "Stat Date"
          },
          "spend": {
            "type": "string",
            "format": "decimal",
            "title": "Spend"
          },
          "revenue": {
            "type": "string",
            "format": "decimal",
            "title": "Revenue"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "fx_rate_to_base": {
            "type": "string",
            "format": "decimal",
            "title": "Fx Rate To Base"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          }
        },
        "type": "object",
        "required": [
          "campaign_run_id",
          "stat_date",
          "currency",
          "fx_rate_to_base",
          "source",
          "id",
          "company_id"
        ],
        "title": "CampaignRunStatOut"
      },
      "CampaignRunUpdate": {
        "properties": {
          "campaign_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "ad_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Account Id"
          },
          "buyer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "CampaignRunUpdate"
      },
...
      "CommitImportRequest": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "column_mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Column Mapping"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "column_mapping"
        ],
        "title": "CommitImportRequest"
      },
      "CommitResult": {
        "properties": {
          "imported": {
            "type": "integer",
            "title": "Imported"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates"
          }
        },
        "type": "object",
        "required": [
          "imported",
          "duplicates"
        ],
        "title": "CommitResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "preview": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Preview"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "columns",
          "row_count",
          "preview"
        ],
        "title": "UploadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}